Add files via upload
A denoising autoencoder is only a minor modification of a "vanilla" autoencoder. However, as opposed to simply reconstructing its input x, it seeks to reconstruct it after a noise function has corrupted it. Doing so allows the model to learn to recover from the noise introduced by the noise function- or "denoising".

For this task we will use a noise function that removes one or two quadrants of the image.
To complete the task, we are expected to:
1. Implement a denoising autoencoder with mean squared error loss for the Fashion-MNIST data.
2. Explote model variants with different layer sizes, depths etc. to find what works well on the validation set and describe the process through which you arrive at your final model.
3. Train your final model to convergence on the training set using an optimisation algorithm of your choice.
4. Provide a plot of the loss on the training set using an optimisation algorithm of your choice.
5. Provide a selection of 32 images in their original, "noisy" and "denoised" form from the test set and comment on what the model has been able to learn and what it appears to find challengeing.