Create README.md
A denoising autoencoder is a minor modification of a "vanilla" autoencoder. However, as opposed to simply reconstructing its input x, it seeks to reconstruct it after a noise function has corrupted it. Doing so allows the model to learn to recover from the noise introduced by the noise function.

For this task we will use a noise function that removes on or two quadrants of the image.
To complete this task you are expected to:
1. Implement a denoising autoencoder with mean squared error loss for the Fashion-MNIST dataset.
2. Explore model variants with different layer sizes, depths, etc. to find what works well on the validation set and describe the process through which you arrive at your final model.
3. Train your final model to convergence on the training set using an optimisation algorithm of your choice.
4. Provide a plot of the loss on the training set and validation set for each epoch of training.
5. Provide a selection of 32 images in ther original, "noisy" and denoised form from the test set and comment on what the model has been able to learn and what it appears to find challenging.